+++ spec/features/comments_feature_spec.rb
@@ -1,13 +1,9 @@
-require 'rails_helper'
-
 feature 'comments' do
   before { Post.create description: 'Awesome latte' }
 
   scenario 'allows user to leave a comment' do
     sign_up('test@test.com', 'testtest', 'testtest')
-    click_link 'Write a comment'
-    fill_in 'Thoughts', with: 'Great picture'
-    click_button 'Comment'
+    write_comment('Great picture')
     expect(page).to have_content 'Great picture'
     expect(current_path).to eq '/posts'
   end
@@ -20,19 +16,9 @@ feature 'comments' do
 
   scenario 'an invalid comment' do
     sign_up('test@test.com', 'testtest', 'testtest')
-    click_link 'Write a comment'
-    fill_in 'Thoughts', with: 'XY'
-    click_button 'Comment'
+    write_comment('XY')
     expect(page).not_to have_css 'li', text: 'XY'
     expect(page).to have_content 'error'
   end
 
-end
-
-def sign_up(email, password, password_confirmation)
-  visit '/users/sign_up'
-  fill_in 'Email', with: email
-  fill_in 'Password', with: password
-  fill_in 'Password confirmation', with: password_confirmation
-  click_button 'Sign up'
 end
